requirements.txt
boto3==1.35.30
botocore==1.35.30
Flask==3.0.3
Flask_Login==0.6.3
Jinja2==3.1.2
Pillow==10.4.0
PyJWT==2.9.0  # Keeping the latest version
pymongo==4.10.0
python_dateutil==2.8.2
PyYAML==6.0.2  # Keeping the latest version
Requests==2.32.3
Werkzeug==3.0.4
